Flatten the file hierarchy
