--- packages/editor/src/plugins/paths/index.ts
@@ -1,22 +1,0 @@
-import {
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin
-} from '@jupyterlab/application';
-
-import { App } from '../../app/app';
-
-/**
- * The default paths.
- */
-const paths: JupyterFrontEndPlugin<JupyterFrontEnd.IPaths> = {
-  id: 'jupyterlab-app-template:paths',
-  activate: (
-    app: JupyterFrontEnd<JupyterFrontEnd.IShell>
-  ): JupyterFrontEnd.IPaths => {
-    return (app as App).paths;
-  },
-  autoStart: true,
-  provides: JupyterFrontEnd.IPaths
-};
-
-export default paths;
